move snapshot functions into missionSet
